requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
coverage==7.1.0
exceptiongroup==1.1.0
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-cov==4.0.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.0
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.0
